\encoding{UTF-8}
\name{beatenberg}
\alias{beatenberg}
\title{Beatenberg}
\description{
  Preprocessed pairs of temperatures in unit Fréchet scale from
  Beatenberg forest, registered under forest cover and in the open
  field. Preprocessing is conducted as described in Ferrez et al
  (2011).
}
\usage{beatenberg}
\format{
  The \code{beatenberg} data frame has 2839 rows and 2 columns.
}
\references{
  Ferrez, J., A. C. Davison, and Rebetez., M. (2011) Extreme
  temperature analysis under forest cover compared to an open field.
  \emph{Agricultural and Forest Meteorology}, 151, 992--1001.
}
\keyword{datasets}