\encoding{UTF-8}
\name{extremis-package}
\alias{extremis-package}
\alias{extremis}
\docType{package}
\title{Statistical Modeling of Extreme Events}
\description{
  \if{html}{
    \figure{logo.png}{options: alt='logo' width='120'}
  }  
  \if{latex}{\figure{logo.png}{options: width=3cm}}
  
  \pkg{extremis} is an add-on tool for \R that conducts inference for
  statistical models of extreme values. For a complete list of
  functions, datasets, and documentation, type \code{help.start()} and
  follow the link to \pkg{extremis} on the Package Index.
}
\author{Miguel de Carvalho; School of Mathematics, University of Edinburgh.}
\seealso{
  Useful links:
  \itemize{
    \item \url{https://www.maths.ed.ac.uk/~mdecarv/publications.html}
  }
}
\section{Funding}{
  Fundação para a Ciência e a Tecnologia (Portuguese NSF) grants:
  \itemize{
    \item PTDC/MAT-STA/28649/2017.
    \item UID/MAT/00006/2019.
  }
}
\keyword{extremis}
